moves getTransactionNotes to utils
